level design beginning
adding new assets and redesigning levels that will make use of the mechanics in the game. started with the first two levels. Level one will be simple for the player as it always spawns first but the rest should not have difficulty changes as there is no way to make them spawn at certain points and that is definitely not in the scope