Enemy Spawner
Began the enemy spawner. Enemies will spawn at each wave. Now I have to control this with the room manager and doors in order to allow the player to kill a certain amount of waves until the doors are unlocked when they have all been killed. Then the player will go to one of the next rooms and repeat this process