wave manager changes
We can now see what player the room is in. With this information we are able to add tags to the spawners in the same room as the player that will allow them to function. This will fix the problem with the enemies spawning in the wrong room and we will be able to play through the whole game.