room tag extended
Can see when the player goes into every room. Need to figure out the proper way of adding and removing the tags or setting the active tags variable. This is what is stopping us from moving through the levels so once this small thing is fixed we can move on.