Target Lock
The player can now lock onto an enemy infront of them. This will make the ranged combat feel better and less difficult as there was no way to tell where the spell was going to go before this point. Need to implement a way for the player to be able to switch between the enemies infront of them as they may not want to attack the one that has been chosen by the game.