New lightning ability
New ultimate ability for the player where they are able to spawn a load of lightning strikes from over head. Going to add another one that is very similar to it for the swords. Over the next week we are going to shift our focus more to the player abilities and getting all of them working together properly with the animations flowing into eachother