Barebones ability select screen
Ability select screen with just the functionality to give the player a random ability. Now have to come up with a way to make the name, image and description as well as the stats of each of the abilities displayed within the buttons. Will make an ability button to achieve this as there will be a need for variables