Battle music and lighting changes
To do more changes in the levels and create better lighting in them all. Then I will move all the enemy attacks to a component and start working on the boss level and boss enemy for the end of the game. Also need to start working on the HUB area