many changes in levels
created better lighting and added more windows so everything is now much easier to see. Also have started adding lore elements into the levels that will tease the boss fight at the end. Also fixed some things with the cooldown. Next will be to make the cooldown work for the one insance of the ability and to make the spells fire in a better way