level design things
beefed up room 5 with lights and stuff so it is more playable. Need to make more decoration and make sure that it is working properly with the rest of the levels. All the levels are now also linked up with the new level. Added more lights to the rest of the levels as well as the new one