New player blocking
Added player block to the player as well as making it function a lot better and feel better to use. May do this to the teleport as well as it might be a good idea to have the player be able to use these things from the start so they are able to get used to them.