bug fixes and some improments
broke ai when making the boss so I have now fixed all of that. Player is actually shooting at the centre of the screen we just need that to work when the player is targeting. New particle system for block ability. When the boss dies the player is sent to the win screen (this is something i want to make into a cutscene but probably wont have time)