Created first imbue ability
Player can now imbue their staff with fire to add damage. Will make it so there is a chance to burn the enemy when hitting the enemy with the staff after this ability has been activated. Furthermore,  I will be making more abilities like this for other element types as well as a stun, push back and stuck in the ground ability for the player to add more variety to the abilities.