Boss debug
Fixed issue where boss was getting stuck in teleporting state. Added boss health bar back to the HUD and set the health of the boss to a good level for when the player is fighting them as before if the player was at max level they would be able to kill the enemy very fast.