enemy spawn fix
enemy spawning is better and has new particles. have also fixed infinite loop problem. Need to fix ai sometimes not attacking player and the mage enemy being spawned in the wrong place still in room 3 not happening in other rooms so definetly has something to do with the way the room is set up rather than the code in the room manager