Adding new room and changing room 3
room 3 is still causing issues with the enemies spawning in so I am going to redo it. Also going to add another room because even i am getting bored of running around in the same 4. Going to also invert the levels as new rooms to have as much variety as possible in the remaining time frame