Save game begin, another package
packaged the game again for more playtesting with all the bugs fixed from last time. Have started to create the infrastructure for the save game functionality. Will have to think really hard how i am going to do this as the way the game is at the minute, it will reset everything about the player when they spawn into the game, making the save obselte. Of course this is what we want in the case of the player dying, but not when the player just wants to have a break from the game.