Save game added
Player can now save their progress if they want to end a run early but dont want to lose where they have got to. Furthermore, when the player dies the save is deleted, keeping the roguelike genre going. This needs to be improved as it is still quite janky and we also need to work on saving the rest of the things that we need for example more things about the player such as the abilities they have as well as saving the enemies and the room manager as well so we can continue in any room with enemies still there as if you were to save and come back the enemies in that room will not spawn.