Added dialogue for boss
dialogue for when the player walks into the boss room as well as more work on the save game functionality. Having a problem with the health and mana being displayed properly as when the player loads back in, when they have taken damage in their save, it just shows a full health bar no matter what. Same with the mana as well.