Save is working but a lot of issues
I have got the save working the way that I want it apart from the fact that when you die you can still load your save even when its been deleted as I have to create a new object when loading and somehow the new object is the old one or something which makes no sense. As well as this none of the player attributes are being saved to the HUD. For some reason the players health is lower but still looks full on the health bar.