Took away save and fixed boss crashing the game when hit
The save system is gone because there are too many issues with it and I have to submit in the next couple days. Have also fixed an issue with the boss fight where hitting the boss would crash the game. This was due to the fact that we were using logic that was specific to the other enemies and it was breaking the boss since he uses a different health bar to the other enemies. Just basically made the boss enemy use its own begin play rather than using the parent one and had it do all the things it needed to do in order to still work as expected.