UI changes
Have started the process of making a more in depth UI such as making the start menu work better as well as adding credits and a link to discord to provide feedback for the game. I have also started to make a more comprehensive options menu with a bar at the top allowing the player to choose different categories to change including general, audio, graphics and controls.