Added new characters and staffs
New characters will be unlockable as the player is able to defeat more bosses in the game. The player will also have the chance to be able to get different staffs which will have different stats and properties. For example the fire staff will have a chance to apply the burn status effect. This will also be useful for bosses who use a staff as they will be able to inflict this on the player also.